pull json code into android module
