+++ segment-android/src/main/java/com/f2prateek/segment/android/MessageRetrofitConverter.java
@@ -2,7 +2,6 @@ package com.f2prateek.segment.android;
 
 import android.util.JsonWriter;
 import com.f2prateek.segment.model.Batch;
-import com.f2prateek.segment.model.JsonAdapter;
 import java.io.IOException;
 import java.io.OutputStreamWriter;
 import okhttp3.MediaType;
@@ -16,7 +15,7 @@ final class MessageRetrofitConverter implements Converter<Batch, RequestBody> {
   @Override public RequestBody convert(Batch batch) throws IOException {
     Buffer buffer = new Buffer();
     JsonWriter writer = new JsonWriter(new OutputStreamWriter(buffer.outputStream()));
-    JsonAdapter.toJson(writer, batch);
+    JsonUtils.toJson(writer, batch);
     writer.close();
     return RequestBody.create(MEDIA_TYPE, buffer.readByteString());
   }
