+++ segment-model/src/main/java/com/f2prateek/segment/model/TrackMessage.java
@@ -18,12 +18,12 @@ import static com.f2prateek.segment.model.Utils.isNullOrEmpty;
  * @see <a href="https://segment.com/docs/spec/track/">Track</a>
  */
 public final class TrackMessage extends Message {
-  private final String event;
-  private final Map<String, Object> properties;
+  private @NonNull final String event;
+  private @Nullable final Map<String, Object> properties;
 
   @Private TrackMessage(Type type, String messageId, Date timestamp, Map<String, Object> context,
-      Map<String, Object> integrations, String userId, String anonymousId, String event,
-      Map<String, Object> properties) {
+      Map<String, Object> integrations, String userId, String anonymousId, @NonNull String event,
+      @Nullable Map<String, Object> properties) {
     super(type, messageId, timestamp, context, integrations, userId, anonymousId);
     this.event = event;
     this.properties = properties;
